+++ src/backend/profileFiltersHelper.juridique.test.ts
@@ -121,6 +121,9 @@ function getBudgetFinanceJuridique(): EntitéJuridiqueBudgetFinance {
     resultatNetComptable: 30,
     ratioDependanceFinanciere: 31,
     tauxDeCafNetSan: 32,
+    fondsDeRoulement: 33,
+    besoinFondsDeRoulement: 34,
+    tresorerie: 35,
   }
 }
 
@@ -203,6 +206,9 @@ function getBudgetFinanceProfile() {
     ratioDépendanceFinancière: "ok",
     tauxDeCafNette: "ok",
     allocationDeRessources: "ok",
+    fondsDeRoulement: "ok",
+    besoinFondsDeRoulement: "ok",
+    tresorerie: "ok",
   }
 }
 
@@ -686,6 +692,63 @@ describe("Filtre les informations de budget et finance des etablissement juridiq
     let entiteJuridiqueResult = getFullEntiteJuridique();
     const profile = getFullProfile();
     profile.budgetEtFinance.tauxDeCafNette = "Ko";
+
+    // When
+    entiteJuridiqueResult = filterEntiteJuridique(entiteJuridiqueResult, profile);
+
+    // Then
+    expect(entiteJuridiqueResult.budgetFinance).toEqual([expectedBudgetAndFinance]);
+  })
+
+  it("retire les info sur les fonds de roulement si il n’y a pas les droits", () => {
+    // Given
+    const rawBudgetAndFinance = getBudgetFinanceJuridique();
+    const expectedBudgetAndFinance = {
+      ...rawBudgetAndFinance,
+      fondsDeRoulement: "",
+    }
+
+    let entiteJuridiqueResult = getFullEntiteJuridique();
+    const profile = getFullProfile();
+    profile.budgetEtFinance.fondsDeRoulement = "Ko";
+
+    // When
+    entiteJuridiqueResult = filterEntiteJuridique(entiteJuridiqueResult, profile);
+
+    // Then
+    expect(entiteJuridiqueResult.budgetFinance).toEqual([expectedBudgetAndFinance]);
+  })
+
+  it("retire les info sur le besoin en fonds de roulement si il n’y a pas les droits", () => {
+    // Given
+    const rawBudgetAndFinance = getBudgetFinanceJuridique();
+    const expectedBudgetAndFinance = {
+      ...rawBudgetAndFinance,
+      besoinFondsDeRoulement: "",
+    }
+
+    let entiteJuridiqueResult = getFullEntiteJuridique();
+    const profile = getFullProfile();
+    profile.budgetEtFinance.besoinFondsDeRoulement = "Ko";
+
+    // When
+    entiteJuridiqueResult = filterEntiteJuridique(entiteJuridiqueResult, profile);
+
+    // Then
+    expect(entiteJuridiqueResult.budgetFinance).toEqual([expectedBudgetAndFinance]);
+  })
+
+  it("retire les info sur la trésorerie si il n’y a pas les droits", () => {
+    // Given
+    const rawBudgetAndFinance = getBudgetFinanceJuridique();
+    const expectedBudgetAndFinance = {
+      ...rawBudgetAndFinance,
+      tresorerie: "",
+    }
+
+    let entiteJuridiqueResult = getFullEntiteJuridique();
+    const profile = getFullProfile();
+    profile.budgetEtFinance.tresorerie = "Ko";
 
     // When
     entiteJuridiqueResult = filterEntiteJuridique(entiteJuridiqueResult, profile);