+++ src/frontend/ui/commun/StringFormater.ts
@@ -68,7 +68,7 @@ const StringFormater = {
   supprimeLesAccents(str: string) {
     return str
       ?.normalize("NFD")
-      .replace(/[\u0300-\u036f]/g, "")
+      .replaceAll(/[\u0300-\u036f]/g, "")
       .toLowerCase()
       .trim();
   }